+++ fractalistic/utils.py
@@ -1,5 +1,7 @@
 import os
 from typing import Any
+from .fractals import *
+from .colors import *
 
 SRC_DIR = os.path.dirname(os.path.abspath(__file__))
 
@@ -12,3 +14,9 @@ class Vec():
 
     def __str__(self):
         return f"Vec({self.x}, {self.y})"
+
+def get_fractal_index_from_name(name: str):
+    return {frac.__name__: i for i, frac in enumerate(fractal_list)}[name]
+
+def get_color_index_from_name(name: str):
+    return {color.__name__: i for i, color in enumerate(color_renderers)}[name]